# Speed up queries for deliverable dashboard.