For seed data:
Loading time before: 1279ms (1170.8ms Views, 69.4ms ActiveRecord)
Loading time now: 997ms (927.5ms Views, 31.3ms ActiveRecord)

Not a lot, but it's something. The `assignments/_deliverable_dashboard` partial on its own goes from 450.4ms to 239.3ms.